requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-webpack-loader==0.5.0
djangorestframework==3.7.3
djangorestframework-simplejwt==3.2
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pygments==2.2.0
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
